Creating module to update connection template in networks
