+++ libraries/resource_providers/api200/connection_template_provider.rb
@@ -39,6 +39,29 @@ module OneviewCookbook
         @item['bandwidth'] = resource_named(:ConnectionTemplate).get_default(@item.client)['bandwidth']
         create_or_update
       end
+
+      # Contains helper methods to include operations with networks
+      module ConnectionTemplateHelper
+        def update_connection_template(bandwidth)
+          bandwidth = convert_keys(bandwidth, :to_s)
+          connection_template = load_resource(:ConnectionTemplate, uri: @item['connectionTemplateUri'])
+          if connection_template.like? bandwidth
+            Chef::Log.info("#{@resource_name} '#{@name}' connection template is up to date")
+          else
+            diff = get_diff(connection_template, bandwidth)
+            Chef::Log.info "Updating #{@resource_name} '#{@name}' connection template settings#{diff}"
+            @context.converge_by "Update #{@resource_name} '#{@name}' connection template settings" do
+              connection_template.update(bandwidth)
+            end
+          end
+        end
+
+        def reset_connection_template
+          @item.retrieve! || raise("#{@resource_name} '#{@name}' not found!")
+          klass = resource_named(:ConnectionTemplate)
+          update_connection_template(bandwidth: klass.get_default(@item.client)['bandwidth'])
+        end
+      end
     end
   end
 end
