+++ libraries/resource_providers/api200/ethernet_network_provider.rb
@@ -13,6 +13,7 @@ module OneviewCookbook
   module API200
     # EthernetNetworkProvider
     class EthernetNetworkProvider < ResourceProvider
+      include API200::ConnectionTemplateProvider::ConnectionTemplateHelper
       def create_or_update
         bandwidth = @item.data.delete('bandwidth')
         super
@@ -23,26 +24,6 @@ module OneviewCookbook
         bandwidth = @item.data.delete('bandwidth')
         created = super
         update_connection_template(bandwidth: bandwidth) if bandwidth && created
-      end
-
-      def update_connection_template(bandwidth)
-        bandwidth = convert_keys(bandwidth, :to_s)
-        connection_template = load_resource(:ConnectionTemplate, uri: @item['connectionTemplateUri'])
-        if connection_template.like? bandwidth
-          Chef::Log.info("#{@resource_name} '#{@name}' connection template is up to date")
-        else
-          diff = get_diff(connection_template, bandwidth)
-          Chef::Log.info "Updating #{@resource_name} '#{@name}' connection template settings#{diff}"
-          @context.converge_by "Update #{@resource_name} '#{@name}' connection template settings" do
-            connection_template.update(bandwidth)
-          end
-        end
-      end
-
-      def reset_connection_template
-        @item.retrieve! || raise("#{@resource_name} '#{@name}' not found!")
-        klass = resource_named(:ConnectionTemplate)
-        update_connection_template(bandwidth: klass.get_default(@item.client)['bandwidth'])
       end
     end
   end
